Rename some jira loaders
