+++ jira/actions/issues/updateField.ts
@@ -1,4 +1,4 @@
-import type { AppContext } from "../mod.ts";
+import type { AppContext } from "../../mod.ts";
 
 export interface Props {
   /**
